refactor: handle post
